Full refactor cleanup

+++ tools/tests/perft.h
@@ -2,11 +2,13 @@
 #define PERFT_H
 
 #include <map>
+#include <cstdint>
 
+#include "utils/Types.h"
 
 extern int maximum;
 
-int perftSearch(Board &boardInstance, int maxDepth = maximum);
+uint64_t perftSearch(Board &boardInstance, int maxDepth = maximum);
 extern std::map<std::string,int> moveBreakDown;
